+++ lib/onboarding_overlay.dart
@@ -1,7 +1,5 @@
 library onboarding_overlay;
 
-export 'src/constants.dart'
-    show kOverlayRatio, kLabelBoxWidthRatio, kLabelBoxHeightRatio;
 export 'src/label_painter.dart' show LabelPainter, ArrowPosition;
 export 'src/onboarding_widget.dart' show Onboarding, OnboardingState;
 export 'src/overlay_painter.dart' show OverlayPainter;