+++ lib/src/constants.dart
@@ -4,3 +4,5 @@ const double kLabelBoxWidthRatioLarge = 0.8;
 const double kLabelBoxHeightRatio = 0.45;
 const double kArrowHeight = 16.0;
 const double kSpace = 4.0;
+const int kTitleMaxLines = 2;
+const int kBodyMaxLines = 5;